Conversion: Add Val back
Fix topics to use UnityAction.
Add Convert<DelegateA->DelegateB> to support Action->DelegateAction.
Modify Val to have its own UnityEvent; its created topic gets any additional subscribers from the event. TODO: subscribe/unsubscribe its event to a permanently associated topic; I don't think there's a good time to do that unfortunately (_maybe_ I could do it ondeserialize, but that feels really bad wrt OnEnable/OnDisable storms...).

- Modify Camera & math a little.

+++ BDUtil/Collections/EnumArray.cs
@@ -36,14 +36,14 @@ namespace BDUtil
         public bool RemoveKey(TKey key) { Add(key, default); return true; }
         public bool RemoveKey(TKey key, out TValue value)
         {
-            int index = Enums<TKey>.GetOffset(key);
+            int index = (int)Enums<TKey>.GetOffset(key);
             value = Data[index];
             Data[index] = default;
             return true;
         }
         public bool TryGetValue(TKey key, out TValue value)
         {
-            int index = Enums<TKey>.GetOffset(key);
+            int index = (int)Enums<TKey>.GetOffset(key);
             value = Data[index];
             return true;
         }
